Moved classes back to utilities
