+++ include/aspect/compositional_initial_conditions/ascii_data.h
@@ -23,7 +23,6 @@
 #define __aspect__compositional_initial_conditions_ascii_data_h
 
 #include <aspect/compositional_initial_conditions/interface.h>
-#include <aspect/initial_conditions/ascii_data.h>
 
 #include <aspect/simulator_access.h>
 #include <aspect/utilities.h>
@@ -42,7 +41,7 @@ namespace aspect
      * @ingroup CompositionalInitialConditionsModels
      */
     template <int dim>
-    class AsciiData : public Interface<dim>, public InitialConditions::internal::AsciiDataInitial<dim>
+    class AsciiData : public Interface<dim>, public Utilities::AsciiDataInitial<dim>
     {
       public:
         /**
